Ability Videos and Icons
All videos and all icons for every ability for every hero has been scraped from the website and is integrated into the project.
Videos may still need settings setup.
The data portion of this inside the code is fully working on the panel side of things.
The data has been fully setup for Ana to show that it's all working.
The data for all other heroes is stubbed out, but the URLs are incomplete and I will need to copy and paste the file names into Main as my next task.